refactor: centralize layout logic into a new reusable ShowPageLayout component and remove redundant scroll state management from individual pages.
