+++ web/src/modules/tendering/costing-sheets/CostingSheetShowPage.tsx
@@ -9,7 +9,7 @@ import { DocumentChecklistSection } from "@/modules/tendering/checklists/compone
 import { CostingSheetSection } from "./components/CostingSheetView";
 import { BidSubmissionSection } from "@/modules/tendering/bid-submissions/components/BidSubmissionView";
 import { paths } from "@/app/routes/paths";
-import { ShowPageLayout } from "@/modules/tendering/components/ShowPageLayout";
+import { ShowPageLayout } from "@/components/layout/ShowPageLayout";
 
 // CostingSheet page shows all steps EXCEPT "result"
 const STEP_IDS = ["tender-details", "physical-docs", "rfq", "emd-fees", "checklist", "costing", "bid"];
@@ -23,7 +23,6 @@ export default function CostingSheetShowPage() {
     const steps = allSteps.filter((s) => STEP_IDS.includes(s.id));
 
     const [expandedSections, setExpandedSections] = useState<Set<string>>(new Set(["costing"]));
-    const [activeSection, setActiveSection] = useState("costing");
 
     const toggleSection = useCallback((id: string) => {
         setExpandedSections((prev) => {
@@ -34,10 +33,6 @@ export default function CostingSheetShowPage() {
     }, []);
     const expandAll   = useCallback(() => setExpandedSections(new Set(steps.map((s) => s.id))), [steps]);
     const collapseAll = useCallback(() => setExpandedSections(new Set()), []);
-    const jumpToSection = useCallback((id: string) => {
-        setActiveSection(id);
-        document.getElementById(`section-${id}`)?.scrollIntoView({ behavior: "smooth", block: "start" });
-    }, []);
 
     const renderSectionContent = useCallback((stepId: string) => {
         switch (stepId) {
@@ -55,9 +50,6 @@ export default function CostingSheetShowPage() {
     return (
         <ShowPageLayout
             steps={steps}
-            activeSection={activeSection}
-            onJump={jumpToSection}
-            onSectionVisible={setActiveSection}
             expandedSections={expandedSections}
             onToggleSection={toggleSection}
             onExpandAll={expandAll}
